requirements.txt
requests==2.31.0
feedparser==6.0.10
schedule==1.2.1
python-telegram-bot==20.7
anthropic==0.18.1
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.1.0
google-generativeai==0.3.2 
